requirements.txt
setuptools==69.1.1
python-dotenv~=1.0.1
mpmath~=1.3.0